crate: Refactor census2010 module into its own file
